backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
djangorestframework-simplejwt>=5.3
channels>=4.0
daphne>=4.0
Pillow>=10.0
python-dotenv>=1.0
torch>=2.0
torchvision>=0.15
scikit-learn>=1.3
numpy>=1.24
